CLOUD-57128 Completely remove old flow implementation, phase 2
