+++ core/src/main/java/com/sequenceiq/cloudbreak/core/bootstrap/service/container/ContainerBootstrapApiCheckerTask.java
@@ -2,7 +2,7 @@ package com.sequenceiq.cloudbreak.core.bootstrap.service.container;
 
 import org.springframework.stereotype.Component;
 
-import com.sequenceiq.cloudbreak.core.flow.context.ContainerBootstrapApiContext;
+import com.sequenceiq.cloudbreak.core.bootstrap.service.container.context.ContainerBootstrapApiContext;
 import com.sequenceiq.cloudbreak.service.CloudbreakServiceException;
 import com.sequenceiq.cloudbreak.service.StackBasedStatusCheckerTask;
