+++ core/src/main/java/com/sequenceiq/cloudbreak/core/bootstrap/service/host/HostBootstrapApiCheckerTask.java
@@ -2,7 +2,7 @@ package com.sequenceiq.cloudbreak.core.bootstrap.service.host;
 
 import org.springframework.stereotype.Component;
 
-import com.sequenceiq.cloudbreak.core.flow.context.HostBootstrapApiContext;
+import com.sequenceiq.cloudbreak.core.bootstrap.service.host.context.HostBootstrapApiContext;
 import com.sequenceiq.cloudbreak.service.CloudbreakServiceException;
 import com.sequenceiq.cloudbreak.service.StackBasedStatusCheckerTask;
