+++ core/src/main/java/com/sequenceiq/cloudbreak/core/bootstrap/service/host/HostClusterAvailabilityCheckerTask.java
@@ -6,7 +6,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
-import com.sequenceiq.cloudbreak.core.flow.context.HostOrchestratorClusterContext;
+import com.sequenceiq.cloudbreak.core.bootstrap.service.host.context.HostOrchestratorClusterContext;
 import com.sequenceiq.cloudbreak.service.StackBasedStatusCheckerTask;
 
 @Component
