+++ core/src/main/java/com/sequenceiq/cloudbreak/core/bootstrap/service/host/context/HostOrchestratorClusterContext.java
@@ -1,29 +1,29 @@
-package com.sequenceiq.cloudbreak.core.flow.context;
+package com.sequenceiq.cloudbreak.core.bootstrap.service.host.context;
 
 import java.util.HashSet;
 import java.util.Set;
 
 import com.sequenceiq.cloudbreak.domain.Stack;
-import com.sequenceiq.cloudbreak.orchestrator.container.ContainerOrchestrator;
+import com.sequenceiq.cloudbreak.orchestrator.host.HostOrchestrator;
 import com.sequenceiq.cloudbreak.orchestrator.model.GatewayConfig;
 import com.sequenceiq.cloudbreak.orchestrator.model.Node;
 import com.sequenceiq.cloudbreak.service.StackContext;
 
-public class ContainerOrchestratorClusterContext extends StackContext {
+public class HostOrchestratorClusterContext extends StackContext {
 
-    private ContainerOrchestrator containerOrchestrator;
+    private HostOrchestrator hostOrchestrator;
     private GatewayConfig gatewayConfig;
     private Set<Node> nodes = new HashSet<>();
 
-    public ContainerOrchestratorClusterContext(Stack stack, ContainerOrchestrator containerOrchestrator, GatewayConfig gatewayConfig, Set<Node> nodes) {
+    public HostOrchestratorClusterContext(Stack stack, HostOrchestrator hostOrchestrator, GatewayConfig gatewayConfig, Set<Node> nodes) {
         super(stack);
-        this.containerOrchestrator = containerOrchestrator;
+        this.hostOrchestrator = hostOrchestrator;
         this.gatewayConfig = gatewayConfig;
         this.nodes = nodes;
     }
 
-    public ContainerOrchestrator getContainerOrchestrator() {
-        return containerOrchestrator;
+    public HostOrchestrator getHostOrchestrator() {
+        return hostOrchestrator;
     }
 
     public GatewayConfig getGatewayConfig() {
