--- core/src/main/java/com/sequenceiq/cloudbreak/core/flow/FlowHandler.java
@@ -1,7 +1,0 @@
-package com.sequenceiq.cloudbreak.core.flow;
-
-/**
- * Marker interface for flow handlers.
- */
-public interface FlowHandler {
-}
