+++ core/src/main/java/com/sequenceiq/cloudbreak/core/flow2/service/ErrorHandlerAwareFlowEventFactory.java
@@ -1,4 +1,4 @@
-package com.sequenceiq.cloudbreak.core.flow;
+package com.sequenceiq.cloudbreak.core.flow2.service;
 
 import javax.inject.Inject;
