+++ core/src/main/java/com/sequenceiq/cloudbreak/service/account/ScheduledAccountPreferencesValidator.java
@@ -13,7 +13,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
-import com.sequenceiq.cloudbreak.core.flow.FlowManager;
+import com.sequenceiq.cloudbreak.core.flow2.service.ReactorFlowManager;
 import com.sequenceiq.cloudbreak.domain.AccountPreferences;
 import com.sequenceiq.cloudbreak.domain.Stack;
 import com.sequenceiq.cloudbreak.service.stack.StackService;
@@ -34,7 +34,7 @@ public class ScheduledAccountPreferencesValidator {
     private AccountPreferencesValidator preferencesValidator;
 
     @Inject
-    private FlowManager flowManager;
+    private ReactorFlowManager flowManager;
 
     @Scheduled(cron = EVERY_HOUR_0MIN_0SEC)
     public void validate() {
