+++ core/src/test/java/com/sequenceiq/cloudbreak/service/stack/DefaultStackHostServiceTypeTest.java
@@ -29,7 +29,7 @@ import com.sequenceiq.cloudbreak.api.model.Status;
 import com.sequenceiq.cloudbreak.api.model.StatusRequest;
 import com.sequenceiq.cloudbreak.controller.BadRequestException;
 import com.sequenceiq.cloudbreak.controller.validation.blueprint.BlueprintValidator;
-import com.sequenceiq.cloudbreak.core.flow.FlowManager;
+import com.sequenceiq.cloudbreak.core.flow2.service.ReactorFlowManager;
 import com.sequenceiq.cloudbreak.domain.Cluster;
 import com.sequenceiq.cloudbreak.domain.Credential;
 import com.sequenceiq.cloudbreak.domain.Stack;
@@ -59,7 +59,7 @@ public class DefaultStackHostServiceTypeTest {
     private InstanceMetaDataRepository instanceMetaDataRepository;
 
     @Mock
-    private FlowManager flowManager;
+    private ReactorFlowManager flowManager;
 
     @Mock
     private BlueprintValidator blueprintValidator;
